Fix race condition in TestSqlStateExecution
In testSplitAssignmentBatchSizeGreaterThanMaxPending() the loop with
the assertions in it races against scheduleSourcePartitionedNodes()
which also has a 1sec wait in it before checking again if it can
schedule more work. We fix this by marking the splits as all queued,
rather than running, and also change the timeout to 2 secs in the test
to avoid this race.